+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.XX] - 2025-03-14
+
+### Added
+
+- Products API endpoints:
+  - Count products
+  - Create a product
+  - Update a product by ID
+  - Delete a product by ID
+  - Delete all products
+  - List of product variants
+  - List of product options
+
 ## [1.0.0-beta.12] - 2025-03-14
 
 ### Added